Add mailchimp list select helper
