+++ Dynamic/Types/CheckboxMultipleType.php
@@ -32,7 +32,6 @@ class CheckboxMultipleType implements FormFieldTypeInterface
         return new FormFieldTypeConfiguration(
             'sulu_form.type.checkboxmultiple',
             __DIR__ . '/../../Resources/config/form-fields/field_choices.xml',
-            [],
             'complex'
         );
     }
