+++ Dynamic/Types/CheckboxType.php
@@ -32,7 +32,6 @@ class CheckboxType implements FormFieldTypeInterface
         return new FormFieldTypeConfiguration(
             'sulu_form.type.checkbox',
             __DIR__ . '/../../Resources/config/form-fields/default_field.xml',
-            [],
             'complex'
         );
     }
