+++ Dynamic/Types/DropdownType.php
@@ -33,7 +33,6 @@ class DropdownType implements FormFieldTypeInterface
         return new FormFieldTypeConfiguration(
             'sulu_form.type.dropdown',
             __DIR__ . '/../../Resources/config/form-fields/default_field_extended.xml',
-            [],
             'complex'
         );
     }
