+++ Dynamic/Types/HeadlineType.php
@@ -32,7 +32,6 @@ class HeadlineType implements FormFieldTypeInterface
         return new FormFieldTypeConfiguration(
             'sulu_form.type.headline',
             __DIR__ . '/../../Resources/config/form-fields/field_title.xml',
-            [],
             'additional'
         );
     }
