+++ Dynamic/Types/RecaptchaType.php
@@ -31,7 +31,6 @@ class RecaptchaType implements FormFieldTypeInterface
         return new FormFieldTypeConfiguration(
             'sulu_form.type.recaptcha',
             __DIR__ . '/../../Resources/config/form-fields/default_field.xml',
-            [],
             'special'
         );
     }
