+++ Dynamic/Types/SpacerType.php
@@ -32,7 +32,6 @@ class SpacerType implements FormFieldTypeInterface
         return new FormFieldTypeConfiguration(
             'sulu_form.type.spacer',
             __DIR__ . '/../../Resources/config/form-fields/header_without_required.xml',
-            [],
             'additional'
         );
     }
