+++ Dynamic/Types/TextareaType.php
@@ -32,7 +32,6 @@ class TextareaType implements FormFieldTypeInterface
         return new FormFieldTypeConfiguration(
             'sulu_form.type.textarea',
             __DIR__ . '/../../Resources/config/form-fields/field_example_default.xml',
-            [],
             'basic'
         );
     }
